MAIN

+++ app/Http/Controllers/Dashboard/CategoryController.php
@@ -56,14 +56,98 @@ class CategoryController extends Controller
         return response()->json($categories);
     }
 
-    /**
-     * Show the form for creating a new resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
-    public function create()
+    public function fetch()
     {
-        return view('dashboard.manage-posts.categories.create');
+        $categories = Category::onlyParent()->get();
+        $output = '';
+
+        if ($categories->count() > 0) {
+            $output .= '
+                <table id="tableCategory" class="table table-hover align-items-center overflow-hidden">
+                    <thead>
+                        <tr>
+                            <th>Kategori</th>
+                            <th></th>
+                        </tr>
+                    </thead>
+                    <tbody>
+           ';
+            foreach ($categories as $category) {
+
+                if (file_exists('vendor/dashboard/image/thumbnail-categories/' . $category->thumbnail)) {
+                    $thumbnail = asset('vendor/dashboard/image/thumbnail-categories/' . $category->thumbnail);
+                } else {
+                    $thumbnail = asset('vendor/blog/img/default.png');
+                }
+
+                $output .= '
+                <tr>
+                    <td>
+                        <a href="javascript:void(0)" class="d-flex align-items-center" style="cursor: default">
+                            <img src="' . $thumbnail . '" width="60" class="avatar me-3">
+                            <div class="d-block ml-3">
+                                <span class="fw-bold name-user">' . $category->title . '</span>
+                                <div class="small text-secondary">
+                                        ' . substr($category->description, 0, 20) . '...
+                                </div>
+                            </div>
+                        </a>
+                    </td>
+                    <td>
+                        <div class="btn-group dropleft">
+                            <button class="btn btn-link text-dark dropdown-toggle dropdown-toggle-split m-0 p-0"
+                                data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">
+                                <i class="uil uil-ellipsis-v"></i>
+                            </button>
+                            <div class="dropdown-menu dashboard-dropdown dropdown-menu-start py-1">
+                                <button value="' . $category->id . '" class="show_btn dropdown-item d-flex align-items-center">
+                                    <i class="uil uil-eye text-primary"></i>
+                                    <span class="ml-2">Lihat Kategori</span>
+                                </button>
+                                <button value="' . $category->id . '" class="edit_btn dropdown-item d-flex align-items-center">
+                                    <i class="uil uil-pen text-warning"></i>
+                                    <span class="ml-2">Edit Kategori</span>
+                                </button>
+                                <button value="' . $category->id . '" data-title="' . $category->title . '" class="del_btn dropdown-item d-flex align-items-center">
+                                    <i class="uil uil-trash text-danger"></i>
+                                    <span class="ml-2">Hapus Kategori</span>
+                                </button>
+                            </div>
+                        </div>
+                    </td>
+                </tr>
+            ';
+            }
+            $output .= '</tbody></table>';
+        } else {
+            $output .= '
+            <div class="card-body">
+                <div class="text-center">
+                    <p class="card-text">
+                        <b>Hmm.. sepertinya kategori postingan belum dibuat</b>
+                    </p>
+                </div>
+            </div>
+            ';
+        }
+        echo $output;
+    }
+
+    public function show($id)
+    {
+        $category = Category::onlyParent()->find($id);
+
+        if ($category) {
+            return response()->json([
+                'status' => 200,
+                'data' => $category
+            ]);
+        } else {
+            return response()->json([
+                'status' => 404,
+                'message' => 'Kategori tidak ditemukan'
+            ]);
+        }
     }
 
     /**
@@ -98,46 +182,39 @@ class CategoryController extends Controller
         );
 
         if ($validator->fails()) {
-            if ($request->has('parent_category')) {
-                $request['parent_category'] = Category::select('id', 'title')->find($request->parent_category);
-            }
-            return redirect()->back()->withInput($request->all())->withErrors($validator);
+            return response()->json([
+                'status' => 400,
+                'errors' => $validator->errors()->toArray(),
+            ]);
         } else {
-
             // INSERT DATA
             if ($request->hasFile('thumbnail')) {
-                // $public_path = '../../public_html/blog/';
-                // $path = $public_path . "vendor/dashboard/image/thumbnail-categories/";
                 $path = "vendor/dashboard/image/thumbnail-categories/";
                 $image = $request->file('thumbnail');
                 $newImage = uniqid('CateIMG-', true) . '.' . $image->extension();
                 // Resize Img
                 $resizeImg = Image::make($image->path());
-                $resizeImg->resize(1280, 800)->save($path . '/' . $newImage);
+                $resizeImg->fit(1280, 800)->save($path . '/' . $newImage);
             }
 
             // insert data
             $insert = Category::create([
-                'title'         => $request->title,
-                'slug'          => $request->slug,
-                'description'   => $request->description ?? '',
+                'title'         => $request->input('title'),
+                'slug'          => $request->input('slug'),
+                'description'   => $request->input('description') ?? 'Kategori blog postingan tentang ' . $request->input('title'),
                 'thumbnail'     => $newImage ?? 'default.png',
-                'parent_id'     => $request->parent_category
             ]);
 
             if ($insert) {
-                // Alert success
-                return redirect()->route('categories.index')->with(
-                    'success',
-                    'Kategori baru berhasil disimpan!'
-                );
+                return response()->json([
+                    'status' => 200,
+                    'message' => 'Kategori baru berhasil ditambahkan',
+                ]);
             } else {
-                // Jika gagal
-                if ($request->has('parent_category')) {
-                    $request['parent_category'] = Category::select('id', 'title')->find($request->parent_category);
-                }
-
-                return redirect()->back()->withInput($request->all());
+                return response()->json([
+                    'status' => 400,
+                    'message' => 'Oopss.. Kategori gagal ditambahkan',
+                ]);
             }
         }
     }
@@ -148,9 +225,21 @@ class CategoryController extends Controller
      * @param  \App\Models\Category  $category
      * @return \Illuminate\Http\Response
      */
-    public function edit(Category $category)
+    public function edit($id)
     {
-        return view('dashboard.manage-posts.categories.edit', compact('category'));
+        $category = Category::onlyParent()->find($id);
+
+        if ($category) {
+            return response()->json([
+                'status' => 200,
+                'data' => $category
+            ]);
+        } else {
+            return response()->json([
+                'status' => 404,
+                'message' => 'Kategori tidak ditemukan'
+            ]);
+        }
     }
 
     /**
@@ -160,14 +249,16 @@ class CategoryController extends Controller
      * @param  \App\Models\Category  $category
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, Category $category)
+    public function update(Request $request, $id)
     {
+        $category = Category::onlyParent()->find($id);
+
         // Validation
         $validator = Validator::make(
             $request->all(),
             [
                 'title'         => 'required|alpha_spaces|max:50|min:3',
-                'slug'          => 'unique:categories,slug,' . $category->id,
+                'slug'          => 'unique:categories,slug,' . $id,
                 'thumbnail'     => 'image|mimes:jpg,png,jpeg,gif|max:1024',
                 'description'   => 'nullable|max:400|min:10'
             ],
@@ -186,54 +277,49 @@ class CategoryController extends Controller
         );
 
         if ($validator->fails()) {
-            if ($request->has('parent_category')) {
-                $request['parent_category'] = Category::select('id', 'title')->find($request->parent_category);
-            }
-            return redirect()->back()->withInput($request->all())->withErrors($validator);
+            return response()->json([
+                'status' => 400,
+                'errors' => $validator->errors()->toArray(),
+            ]);
         } else {
-            // UPDATE CATEGORY
-            $category->title        = $request->input('title');
-            $category->slug         = $request->input('slug');
-            $category->description  = $request->input('description') ?? '';
-            $category->parent_id    = $request->input('parent_category');
+            if ($category) {
+                $category->title        = $request->input('title');
+                $category->slug         = $request->input('slug');
+                $category->description  = $request->input('description') ?? 'Kategori blog postingan tentang ' . $category->title;
 
-            if ($request->hasFile('thumbnail')) {
-                // $public_path = '../../public_html/blog/';
-                // $path = $public_path . "vendor/dashboard/image/thumbnail-categories/";
-                $path = "vendor/dashboard/image/thumbnail-categories/";
-                if (File::exists($path . $category->thumbnail)) {
-                    File::delete($path . $category->thumbnail);
-                }
-                $image = $request->file('thumbnail');
-                $newImage = uniqid('CateIMG-', true) . '.' . $image->extension();
-                // Resize Img
-                $resizeImg = Image::make($image->path());
-                $resizeImg->resize(1280, 800)->save($path . '/' . $newImage);
-
-                $category->thumbnail = $newImage;
-            }
-
-            if ($category->isDirty()) {
-
-                // Updated process
-                $cateUpdate = $category->update();
-
-                if ($cateUpdate) {
-                    return redirect()->route('categories.index')->with(
-                        'success',
-                        'Kategori berhasil diperbarui!'
-                    );
-                } else {
-                    if ($request->has('parent_category')) {
-                        $request['parent_category'] = Category::select('id', 'title')->find($request->parent_category);
+                if ($request->hasFile('thumbnail')) {
+                    $path = "vendor/dashboard/image/thumbnail-categories/";
+                    if (File::exists($path . $category->thumbnail)) {
+                        File::delete($path . $category->thumbnail);
                     }
-                    return redirect()->back()->withInput($request->all());
+                    $image = $request->file('thumbnail');
+                    $newImage = uniqid('CateIMG-', true) . '.' . $image->extension();
+                    // Resize Img
+                    $resizeImg = Image::make($image->path());
+                    $resizeImg->fit(1280, 800)->save($path . '/' . $newImage);
+
+                    $category->thumbnail = $newImage;
+                }
+
+                if ($category->isDirty()) {
+
+                    $category->update();
+
+                    return response()->json([
+                        'status' => 200,
+                        'message' => 'Kategori ' . $category->title . ' berhasil diubah'
+                    ]);
+                } else {
+                    return response()->json([
+                        'status' => 200,
+                        'message' => 'Tidak ada perubahan'
+                    ]);
                 }
             } else {
-                return redirect()->route('categories.index')->with(
-                    'success',
-                    'Tidak ada perubahan'
-                );
+                return response()->json([
+                    'status' => 400,
+                    'message' => 'Kategori tidak ditemukan'
+                ]);
             }
         }
     }
@@ -244,39 +330,41 @@ class CategoryController extends Controller
      * @param  \App\Models\Category  $category
      * @return \Illuminate\Http\Response
      */
-    public function destroy(Category $category)
+    public function destroy($id)
     {
-        $usePostCategory = Post::join('category_post', 'posts.id', '=', 'category_post.post_id')
-        ->where('category_post.category_id', $category->id)->get();
+        $category = Category::onlyParent()->find($id);
 
-        if ($usePostCategory->count() >= 1) {
-            Alert::warning(
-                'Warning',
-                "Oops.. kategori " . $category->title . " tidak bisa hapus, karena kategori ini sedang digunakan."
-            )->autoClose(false);
-            return redirect()->back();
-        }
+        if ($category) {
 
-        try {
-            // $public_path = '../../public_html/blog/';
-            // $path = $public_path . "vendor/dashboard/image/thumbnail-categories/";
-            $path = "vendor/dashboard/image/thumbnail-categories/";
-            if (File::exists($path . $category->thumbnail)) {
-                File::delete($path . $category->thumbnail);
+            $usePostCategory = Post::join('category_post', 'posts.id', '=', 'category_post.post_id')
+                ->where('category_post.category_id', $id)->get();
+
+            if ($usePostCategory->count() >= 1) {
+                return response()->json([
+                    'status' => 400,
+                    'message' => 'Oopss.. kategori tidak bisa di hapus, karena kategori ' . $category->title . ' sedang digunakan!'
+                ]);
+            } else {
+
+                $path = "vendor/dashboard/image/thumbnail-categories/";
+                if (File::exists($path . $category->thumbnail)) {
+                    File::delete($path . $category->thumbnail);
+                }
+
+                $category->delete();
+
+                return response()->json([
+                    'status' => 200,
+                    'message' => 'Kategori ' . $category->title . ' berhasil dihapus'
+                ]);
             }
-            // Delete data
-            $category->delete();
-        } catch (\Throwable $th) {
-            Alert::error(
-                'Error',
-                'Failed during data input process.
-                Message: ' . $th->getMessage()
-            )->autoClose(false);
+
+        } else {
+            return response()->json([
+                'status' => 400,
+                'message' => 'Kategori tidak ditemukan'
+            ]);
         }
 
-        return redirect()->back()->with(
-            'success',
-            'Kategori ' . $category->title . ' berhasil dihapus!'
-        );
     }
 }
